Revenue_Mobilization_and_Tracking_System/rms-backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
cffi==1.15.1
click==8.1.3
cryptography==41.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.100.0
greenlet==3.0.3
h11==0.14.0
idna==3.4
passlib==1.7.4
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
pyasn1==0.5.0
pycparser==2.21
pydantic==2.8.2
pydantic-core==2.20.1
pydantic-settings==2.3.4
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
redis==5.0.4
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlalchemy==2.0.15
starlette==0.27.0
typing-extensions==4.6.3
uvicorn==0.22.0
